refactor: move abi encoding and decoding functions to noirc_abi crate
