+++ crates/nargo/src/cli/prove_cmd.rs
@@ -5,11 +5,11 @@ use acvm::FieldElement;
 use acvm::ProofSystemCompiler;
 use acvm::{GateResolution, PartialWitnessGenerator};
 use clap::ArgMatches;
-use noirc_abi::AbiType;
+use noirc_abi::errors::AbiError;
 use noirc_abi::{input_parser::InputValue, Abi};
 use std::path::Path;
 
-use crate::errors::{AbiError, CliError};
+use crate::errors::CliError;
 
 use super::{
     create_named_dir, write_to_file, PROOFS_DIR, PROOF_EXT, PROVER_INPUT_FILE, VERIFIER_INPUT_FILE,
@@ -36,70 +36,6 @@ fn prove(proof_name: &str, show_ssa: bool, allow_warnings: bool) -> Result<(), C
         Ok(_) => Ok(()),
         Err(e) => Err(e),
     }
-}
-
-/// Ordering is important here, which is why we need the ABI to tell us what order to add the elements in
-/// We then need the witness map to get the elements field values.
-fn abi_encode(
-    abi: Abi,
-    witness_map: &BTreeMap<String, InputValue>,
-) -> Result<Vec<FieldElement>, AbiError> {
-    let num_params = abi.num_parameters();
-    let mut encoded_inputs = Vec::new();
-
-    let return_witness_len: u32 = abi
-        .parameters
-        .iter()
-        .find(|x| x.0 == noirc_frontend::hir_def::function::MAIN_RETURN_NAME)
-        .map_or(0, |(_, return_type)| return_type.field_count());
-
-    for (param_name, param_type) in abi.parameters.clone().into_iter() {
-        let value = witness_map
-            .get(&param_name)
-            .ok_or_else(|| AbiError::MissingParam(param_name.clone()))?
-            .clone();
-
-        if !value.matches_abi(&param_type) {
-            return Err(AbiError::TypeMismatch { param_name, param_type, value });
-        }
-
-        if param_name != noirc_frontend::hir_def::function::MAIN_RETURN_NAME
-            || return_witness_len != 1
-            || !matches!(value, InputValue::Undefined)
-        {
-            let encoded_input = input_value_into_witness(value, param_name)?;
-            encoded_inputs.extend(encoded_input);
-        }
-    }
-
-    // Check that no extra witness values have been provided.
-    // Any missing values should be caught by the above for-loop so this only catches extra values.
-    if num_params != witness_map.len() {
-        let param_names = abi.parameter_names();
-        let unexpected_params: Vec<String> =
-            witness_map.keys().filter(|param| !param_names.contains(param)).cloned().collect();
-        return Err(AbiError::UnexpectedParams(unexpected_params));
-    }
-
-    Ok(encoded_inputs)
-}
-
-fn input_value_into_witness(
-    value: InputValue,
-    param_name: String,
-) -> Result<Vec<FieldElement>, AbiError> {
-    let mut encoded_value = Vec::new();
-    match value {
-        InputValue::Field(elem) => encoded_value.push(elem),
-        InputValue::Vec(vec_elem) => encoded_value.extend(vec_elem),
-        InputValue::Struct(object) => {
-            for (name, value) in object {
-                encoded_value.extend(input_value_into_witness(value, name)?)
-            }
-        }
-        InputValue::Undefined => return Err(AbiError::UndefinedInput(param_name)),
-    }
-    Ok(encoded_value)
 }
 
 pub fn compile_circuit_and_witness<P: AsRef<Path>>(
@@ -134,7 +70,7 @@ pub fn parse_and_solve_witness<P: AsRef<Path>>(
     let inputs_vector: Vec<FieldElement> = (0..abi.field_count())
         .map(|index| solved_witness[&Witness::new(index + WITNESS_OFFSET)])
         .collect();
-    let decoded_inputs = abi_decode(abi, &inputs_vector);
+    let decoded_inputs = abi.decode(&inputs_vector);
 
     // Write public inputs into Verifier.toml
     export_public_inputs(&decoded_inputs, abi.clone(), &program_dir)?;
@@ -147,7 +83,7 @@ fn solve_witness(
     witness_map: &BTreeMap<String, InputValue>,
 ) -> Result<BTreeMap<Witness, FieldElement>, CliError> {
     let abi = compiled_program.abi.as_ref().unwrap();
-    let encoded_inputs = abi_encode(abi.clone(), witness_map).map_err(|error| match error {
+    let encoded_inputs = abi.clone().encode(witness_map).map_err(|error| match error {
         AbiError::UndefinedInput(_) => {
             CliError::Generic(format!("{} in the {}.toml file.", error, PROVER_INPUT_FILE))
         }
@@ -179,59 +115,6 @@ fn solve_witness(
     }
 
     Ok(solved_witness)
-}
-
-fn abi_decode(abi: &Abi, encoded_inputs: &Vec<FieldElement>) -> BTreeMap<String, InputValue> {
-    let mut index = 0;
-    let mut decoded_inputs = BTreeMap::new();
-
-    for (param_name, param_type) in &abi.parameters {
-        let (next_index, decoded_value) =
-            read_value_from_witness(index, &encoded_inputs, &param_type);
-
-        decoded_inputs.insert(param_name.to_owned(), decoded_value);
-
-        index = next_index;
-    }
-    decoded_inputs
-}
-
-fn read_value_from_witness(
-    initial_index: usize,
-    encoded_inputs: &Vec<FieldElement>,
-    value_type: &AbiType,
-) -> (usize, InputValue) {
-    let mut index = initial_index;
-
-    let value = match value_type {
-        AbiType::Field(_) | AbiType::Integer { .. } => {
-            let field_element = encoded_inputs[index];
-            index += 1;
-
-            InputValue::Field(field_element)
-        }
-        AbiType::Array { length, .. } => {
-            let field_elements = &encoded_inputs[index..index + (*length as usize)];
-
-            index += *length as usize;
-            InputValue::Vec(field_elements.to_vec())
-        }
-        AbiType::Struct { fields, .. } => {
-            let mut struct_map = BTreeMap::new();
-
-            for (field_key, param_type) in fields {
-                let (next_index, field_value) =
-                    read_value_from_witness(index, encoded_inputs, param_type);
-
-                struct_map.insert(field_key.to_owned(), field_value);
-                index = next_index;
-            }
-
-            InputValue::Struct(struct_map)
-        }
-    };
-
-    (index, value)
 }
 
 fn export_public_inputs<P: AsRef<Path>>(
