+++ crates/noirc_frontend/src/hir_def/function.rs
@@ -1,4 +1,4 @@
-use noirc_abi::{Abi, AbiFEType};
+use noirc_abi::{Abi, AbiFEType, MAIN_RETURN_NAME};
 use noirc_errors::{Location, Span};
 
 use super::expr::{HirBlockExpression, HirExpression, HirIdent};
@@ -12,8 +12,6 @@ use crate::{token::Attribute, FunctionKind};
 /// with a list of statements
 #[derive(Debug, Clone)]
 pub struct HirFunction(ExprId);
-
-pub const MAIN_RETURN_NAME: &str = "return";
 
 impl HirFunction {
     pub fn empty() -> HirFunction {
